segstore: add default unhandled methods to ISegmentStore
This handles common functionalities for all segment stores.
This can be seen as part of the segment store contract.

Anything removing items from a segment store asside from
clear should throw an unsupportedOperationException.

If an implementation decides to support this, that is fine, but
not the expected way the object should behave.

Change-Id: I9d00960ae8826cecc63092245ca7fc447a28defb
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/83305
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>