Output workspace log to console during tests
This patch adds -consoleLog so that we see easily the workspace logs
in the console output. It's very easy to miss problems that were
outputted in the workspace log because you have to find the log in
target/work/data/.log.

One disadvantage is that it always outputs:
"Could not resolve module: org.eclipse.compare.win32"

Change-Id: I021f30713485fc152465bd81f6ac9baf23ef5f16
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/83812
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>