segstore: Add assert on item count for iteration benchmarks
This patch adds asserts on the item count when iterating the segment
store in the benchmark. It makes sure that all iterations actually
iterate over all the elements of the segment and does not give a false
impression of performance if there is a bug in the iterator.

Also remove the System.gc() that are not useful to the test

Change-Id: I87acbb264703f06a166239655f3911697c17126e
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/85724
Reviewed-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@efficios.com>